# apps/web/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "CoverageReport": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/CoverageGapOut"
            },
            "title": "Gaps",
            "type": "array"
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/PropertyCoverage"
            },
            "title": "Properties",
            "type": "array"
          }
        },
        "required": ["as_of", "properties", "gaps"],
        "title": "CoverageReport",
        "type": "object"
      },
      "DeadlineOut": {
        "properties": {
          "citation": {
            "title": "Citation",
            "type": "string"
          },
          "due_on": {
            "format": "date",
            "title": "Due On",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "property_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Label"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "window_opens_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Opens On"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "due_on",
          "window_opens_on",
          "citation",
          "note",
          "property_label"
        ],
        "title": "DeadlineOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The owner's operating platform for real property.",
    "title": "Hestia API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/coverage/jurisdictions": {
      "get": {
        "operationId": "coverage_jurisdictions_coverage_jurisdictions_get",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Coverage Jurisdictions"
      }
    },
    "/deadlines": {
      "get": {
        "operationId": "list_deadlines_deadlines_get",
        "parameters": [
          {
            "in": "query",
            "name": "due_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeadlineOut"
                  },
                  "title": "Response List Deadlines Deadlines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Deadlines"
      }
    },
    "/entities": {
      "get": {
        "operationId": "list_entities_entities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntityOut"
                  },
                  "title": "Response List Entities Entities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Entities"
      },
      "post": {
        "operationId": "create_entity_entities_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Entity"
      }
    },
...
    "/ledger": {
      "get": {
        "operationId": "ledger_register_ledger_get",
        "parameters": [
          {
            "in": "query",
            "name": "property_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Property Id"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "rent",
                    "other_income",
                    "late_fee",
                    "deposit_received",
                    "deposit_returned",
                    "mortgage_interest",
                    "mortgage_principal",
                    "property_tax",
                    "insurance",
                    "repairs",
                    "capital_improvement",
                    "utilities",
                    "management_fee",
                    "hoa",
                    "legal_professional",
                    "advertising",
                    "supplies",
                    "travel",
                    "acquisition_cost",
                    "disposition_cost",
                    "owner_contribution",
                    "owner_distribution"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "occurred_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurred From"
            }
          },
          {
            "in": "query",
            "name": "occurred_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurred To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerRegister"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ledger Register"
      },
      "post": {
        "operationId": "append_ledger_event_ledger_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerEntryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Ledger Event"
      }
    },
    "/ledger/{event_uuid}/reverse": {
      "post": {
        "operationId": "reverse_ledger_event_ledger__event_uuid__reverse_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReversalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReversalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reverse Ledger Event"
      }
    },
    "/properties": {
      "get": {
        "operationId": "list_properties_properties_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PropertySummary"
                  },
                  "title": "Response List Properties Properties Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Properties"
      },
      "post": {
        "operationId": "create_property_properties_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Property"
      }
    },
    "/properties/{property_id}": {
      "get": {
        "operationId": "get_property_properties__property_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Property"
      }
    },
    "/properties/{property_id}/dossier": {
      "get": {
        "operationId": "read_dossier_properties__property_id__dossier_get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DossierView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Dossier"
      },
      "post": {
        "operationId": "assemble_dossier_properties__property_id__dossier_post",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "in": "header",
            "name": "x-actor",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DossierOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assemble Dossier"
      }
    },
...
    "/sweep/deadlines": {
      "post": {
        "operationId": "sweep_deadlines_sweep_deadlines_post",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "in": "header",
            "name": "x-actor",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SweepOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sweep Deadlines"
      }
    }
  }
}